+++ src/test/java/org/opentripplanner/transit/raptor/api/path/PathTest.java
@@ -47,6 +47,7 @@ public class PathTest {
     @Test
     public void numberOfTransfers() {
         assertEquals(2, subject.numberOfTransfers());
+        assertEquals(2, subject.numberOfTransfersExAccessEgress());
     }
 
     @Test