+++ docs/Changelog.md
@@ -34,6 +34,8 @@
 - Per-mode reluctance parameters are added so that itineraries with multiple modes may have varying reluctances. [#3501](https://github.com/opentripplanner/OpenTripPlanner/issues/3501)
 - Add `maxAreaNodes` configuration parameter for changing an area visibility calculation limit [#3501](https://github.com/opentripplanner/OpenTripPlanner/issues/3534)
 - Add bicycle safety report to report API [#3563](https://github.com/opentripplanner/OpenTripPlanner/issues/3563)
+- Optimize Transfers performance issue [#3513](https://github.com/opentripplanner/OpenTripPlanner/issues/3513)
+
 
 ## 2.0.0 (2020-11-27)